chore: move routingContext to hooks/useRouter
