+++ dropbox-hacking-photo-manager-client/src/components/samePageLink.tsx
@@ -1,7 +1,7 @@
 import { RouteState, urlForState } from "dropbox-hacking-photo-manager-shared";
 import * as React from "react";
 
-import { useRouter } from "@/context/routingContext";
+import useRouter from "@/hooks/useRouter";
 
 const samePageLink = ({
   routeState,
